requirements.txt
pillow==11.3.0
python-magic==0.4.27

libheif
